# Conflicts:
#	src/main/java/com/example/financial_case_module_4/controller/CategoryController.java
#	src/main/java/com/example/financial_case_module_4/repository/ITransactionRepository.java
#	src/main/resources/js/category.js
#	src/main/resources/sidebar-01/index.html

+++ src/main/resources/js/category.js
@@ -1,7 +1,7 @@
-let aa = 0;
 function showCategoryFrom() {
     $('#modalAddCategoryForm').modal('show');
 }
+
 function addCategory() {
     let nameCategory = document.getElementById("moneyNameCategory").value;
     let category = {
@@ -9,11 +9,12 @@ function addCategory() {
     }
     $.ajax({
         headers: {
+            Authorization: 'Bearer ' + window.sessionStorage.getItem("TOKEN_KEY"),
             "Accept": 'application/json',
             "Content-Type": 'application/json',
         },
         type: "POST",
-        url: "http://localhost:8000/category",
+        url: "http://localhost:8000/money_categories",
         data: JSON.stringify(category),
         success : function () {
             $('#modalAddCategoryForm').modal('hide');
@@ -25,6 +26,7 @@ function addCategory() {
         }
     })
 }
+
 function showAllCategory() {
     $.ajax({
         headers: {
@@ -33,7 +35,7 @@ function showAllCategory() {
             'Content-Type': 'application/json'
         },
         type: "GET",
-        url: "http://localhost:8000/category" ,
+        url: "http://localhost:8000/money_categories" ,
         success: function (moneyCategory) {
             let content = ``;
             for (let i = 0; i < moneyCategory.length; i++) {
@@ -45,19 +47,20 @@ function showAllCategory() {
         }
     })
 }
-showAllCategory();
+
 function showMoneyCategory(id) {
     $('#modalEditCategoryForm').modal('show');
     $.ajax({
         headers: {
+            Authorization: 'Bearer ' + window.sessionStorage.getItem("TOKEN_KEY"),
             "Accept": 'application/json',
             "Content-Type": 'application/json',
         },
         type: "GET",
-        url: "http://localhost:8000/category/find-by-category-id/" + id,
+        url: "http://localhost:8000/money_categories/find-by-category-id/" + id,
         success : function (data) {
             document.getElementById("showNameCategory").innerHTML = `<div class="modal-header text-center">
-                <h4 class="modal-title w-100 font-weight-bold" style="color: #28b498">Create Category</h4>
+                <h4 class="modal-title w-100 font-weight-bold" style="color: #28b498">Category</h4>
             </div>
             <div class="modal-body mx-3">
                 <div class="md-form mb-5" >
@@ -69,17 +72,19 @@ function showMoneyCategory(id) {
             <input type="hidden" id="idCategory" value="${data.id}">
             <div class="modal-footer d-flex justify-content-center">
                 <button class="btn btn-default" onclick="editFromCategory()">Edit</button>
-                <button class="btn btn-danger" onclick="deleteFromCategory()">Xoa</button>
+                <button class="btn btn-danger" onclick="deleteFromCategory()">Delete</button>
             </div> `;
 
         }
     })
 }
+
 function editFromCategory() {
     editCategory();
     $('#modalEditCategoryForm').modal('hide');
     showAllCategory();
 }
+
 function editCategory() {
     let idEdit = document.getElementById("idCategory").value;
     let name = document.getElementById("moneyEditNameCategory").value;
@@ -89,13 +94,15 @@ function editCategory() {
     console.log(moneyCategory);
     $.ajax({
         headers: {
+            Authorization: 'Bearer ' + window.sessionStorage.getItem("TOKEN_KEY"),
             "Accept": 'application/json',
             "Content-Type": 'application/json',
         },
         type: "PUT",
-        url: "http://localhost:8000/category/" + idEdit ,
+        url: "http://localhost:8000/money_categories/" + idEdit ,
         data: JSON.stringify(moneyCategory),
         success : function () {
+            showAllCategory();
             $("#myModal").modal('show');
             setTimeout(function () {
                 $("#myModal").modal('hide');
@@ -103,17 +110,20 @@ function editCategory() {
         }
     })
 }
+
 function deleteFromCategory() {
     $('#modalEditCategoryForm').modal('hide');
-    let str = "Do You Want delete ?"
-    modalAlert(str)
+    modalAlert(removes)
 }
 
 function deleteCategory() {
     let id = document.getElementById("idCategory").value;
     $.ajax({
+        headers : {
+            Authorization: 'Bearer ' + window.sessionStorage.getItem("TOKEN_KEY"),
+        },
         type: "DELETE",
-        url: "http://localhost:8000/category/" + id,
+        url: "http://localhost:8000/money_categories/" + id,
         success: function () {
             $("#myModal").modal('show');
             setTimeout(function () {
@@ -125,6 +135,8 @@ function deleteCategory() {
     $('#modalAlert').modal('hide')
 }
 
+let removes = "Do You Want delete ?";
+
 function modalAlert(str) {
     $('#modalAlert').modal('show');
     document.getElementById("yesOfNo").innerHTML = `
@@ -133,16 +145,28 @@ function modalAlert(str) {
                 <h4 class="modal-title w-100 font-weight-bold" style="color: #28b498">${str}</h4>
             </div>
             <div class="modal-footer d-flex justify-content-center">
-                <button className="btn btn-default" onclick="yes()">YES</button>
-                <button className="btn btn-danger" onclick="no()">No</button>
+                <button class="btn btn-default" onclick="yes('${str}')">Accpect</button>
+                <button class="btn btn-danger" onclick="no('${str}')">Cancel</button>
             </div>
         </div>`
 }
-function yes() {
-    deleteCategory();
-    $('#modalEditCategoryForm').modal('hide');
+
+function yes(str) {
+    if (str === removes) {
+        deleteCategory();
+        $('#modalEditCategoryForm').modal('hide');
+        $("#myModal").modal('show');
+        setTimeout(function () {
+            $("#myModal").modal('hide');
+        }, 1000);
+    }
 }
-function no() {
-    $('#modalAlert').modal('hide');
-    $('#modalEditCategoryForm').modal('show');
+
+function no(str) {
+    if (str === removes) {
+        resetValue();
+        $('#modalAlert').modal('hide');
+        $('#modalEditCategoryForm').modal('show');
+
+    }
 }
